python/timestep/pyproject.toml
[project]
name = "timestep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.2.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "aiofiles>=23.0.0",
    "openai-agents>=0.3.2",
    "ag-ui-protocol>=0.1.0",
]
